refactor utils and update RandomForestClassifier_v0_1
