+++ pybio/spec/__init__.py
@@ -1,6 +1,6 @@
 from typing import get_type_hints
 
+from pybio.spec.utils import load_model_spec, load_spec, maybe_convert
 from . import nodes
-from .utils import load_spec, load_model_spec
 
 __version__ = get_type_hints(nodes.Spec)["format_version"].__args__
