+++ pybio/spec/utils/common.py
@@ -1,0 +1,3 @@
+from ruamel.yaml import YAML
+
+yaml = YAML(typ="safe")
