+++ specs/models/sklearn/RandomForestClassifier_v0_1.model.yaml
@@ -15,7 +15,7 @@ framework: scikit-learn
 
 source: pybio.sklearn.models.sklearnbased.RandomForestClassifier
 optional_kwargs:
-  c_indices: [null]
+  channel_indices: [null]
   n_estimators: 10
   criterion: gini
   max_depth: null
@@ -25,7 +25,6 @@ optional_kwargs:
   max_features: auto
   max_leaf_nodes: null
   min_impurity_decrease: 0.0
-  min_impurity_split: 1.0e-7
   bootstrap: true
   oob_score: false
   n_jobs: 1
@@ -34,8 +33,8 @@ optional_kwargs:
   warm_start: false
   class_weight: null
 
-test_input: ../test_input.npy
-test_output: ../test_output.npy
+test_input: test_input_raw.npy
+test_output: test_output.npy
 covers: [] # ./nuclei_thumbnail.png
 
 inputs:
@@ -61,8 +60,8 @@ outputs:
 prediction:
     preprocess: []
     weights:
-        source: "https://zenodo.org/record/3446812/files/unet2d_weights.torch"  # todo: give real RF weights
-        hash: {sha256: TOOODOOOTOOODOOOTOOODOOOTOOODOOOTOOODOOOTOOODOOOTOOODOOOTOOODOOO}
+        source: rf_v0.pickle
+        hash: {sha256: abcdefghabcdefghabcdefghabcdefghabcdefghabcdefghabcdefghabcdefgh}
     postprocess: []
     dependencies: conda:../env_numpy.yaml
 
@@ -88,6 +87,12 @@ training:
 
 config:
   future:
-    weights_format: "pickle"
-    timestamp: "2020-10-23T16:00:42+00:00"
-    git_repo: "https://github.com/bioimage-io/python-bioimage-io"
+    0.3.0:
+      weights_format: "pickle"
+      weights:
+        id: default
+        name: toy examle weights
+        description: "Trained as a toy example on a tiny subset of the broad nucleus data. Don't use for any real problem! (Use ilastik pixelclassification instead! ;-))"
+        tags: [toy example]
+      timestamp: "2020-10-23T16:00:42+00:00"
+      git_repo: "https://github.com/bioimage-io/python-bioimage-io/tree/master/specs/models/sklearnbased"
