+++ tests/test_specs/test_format_version_conversion.py
@@ -1,12 +1,11 @@
 from dataclasses import asdict
 
-from pybio.spec import schema
-from pybio.spec.utils import FormatConverter
+from pybio.spec import schema, maybe_convert
 
 
 def test_model_nodes_format_0_1_to_0_3(rf_model_data_v0_1, rf_model_data):
     expected = asdict(schema.Model().load(rf_model_data))
-    converted_data = FormatConverter()(rf_model_data_v0_1)
+    converted_data = maybe_convert(rf_model_data_v0_1)
     actual = asdict(schema.Model().load(converted_data))
 
     # expect converted description
